+++ src/constants/constants.js
@@ -1,6 +1,6 @@
 export const DefaultAppName = 'app';
 export const DefaultTemplate = 'react-app';
-export const TemplateExt = '.mustache';
+export const TemplateExt = 'mustache.';
 export const ManifestoFileName = 'manifesto.json';
 export const DefaultPluralSuffix = 'Collection';
 export const Directive = {