+++ src/templates/react-app/db/sequelize/models/{{modelNamePascal}}.mustache.js
@@ -1,0 +1,14 @@
+import sequelizeModule from 'sequelize';
+import { sequelize } from '../sequelize.js'
+const { DataTypes } = sequelizeModule;
+
+export const {{modelNamePascal}} = sequelize.define('{{modelNamePascal}}', {
+  {{#properties}}
+  {{propertyNameCamel}}: {
+    type: DataTypes.{{sequelizeType}},
+    field: '{{propertyNameSnake}}'
+  },
+  {{/properties}}
+}, {
+  tableName: '{{modelNameSnake}}'
+});