+++ src/templates/react-app/db/sequelize/scripts/mustache.dbSync.js
@@ -1,0 +1,7 @@
+import { sequelize } from '../src/db/sequelize.js';
+{{#models}}
+import "../src/db/models/{{modelNamePascal}}.js";
+{{/models}}
+
+await sequelize.sync({ alter: true });
+console.log('All models were synchronized successfully.');